refactor(news): separate collection from semantic flow
