+++ tests/configs/test_base.py
@@ -50,12 +50,12 @@ class PackageExportTests(unittest.TestCase):
         )
         from quantmind.configs import (
             EarningsFlowCfg,
-            NewsFlowCfg,
+            NewsCollectionCfg,
             PaperFlowCfg,
         )
 
         self.assertTrue(issubclass(PaperFlowCfg, BaseFlowCfgExport))
-        self.assertTrue(issubclass(NewsFlowCfg, BaseFlowCfgExport))
+        self.assertTrue(issubclass(NewsCollectionCfg, BaseFlowCfgExport))
         self.assertTrue(issubclass(EarningsFlowCfg, BaseFlowCfgExport))
         self.assertEqual(BaseInputExport.__name__, "BaseInput")
