+++ tests/preprocess/test_news_types.py
@@ -3,7 +3,7 @@
 import unittest
 from dataclasses import FrozenInstanceError
 
-from quantmind.preprocess.news import NewsArtifact, NewsBatch, NewsFailure
+from quantmind.preprocess import NewsArtifact, NewsBatch, NewsFailure
 
 
 class NewsResultContractTests(unittest.TestCase):
